separate FFICXX.Generate.Type.Class to FFICXX.Generate.Code.Primitive
