+++ sample/STL-generator/Foo.h
