+++ sample/STL-generator/stub.cc
