+++ sample/cxxlib/include/A.h
