+++ sample/cxxlib/include/B.h
