+++ sample/cxxlib/src/A.cpp
