+++ sample/cxxlib/src/B.cpp
