+++ sample/mysample-generator/MySampleGen.hs
