+++ sample/mysample-generator/build.sh
