+++ sample/mysample-generator/use_mysample.hs
