+++ sample/qt-generator/QtGen.hs
