+++ sample/qt-generator/testqt.hs
