+++ sample/snappy-generator/SnappyGen.hs
