+++ sample/snappy-generator/testSnappy.hs
