+++ sample/template-test/stl/Foo.h
