+++ sample/template-test/stl/STL/TH.hs
