+++ sample/template-test/stl/STL/Template.hs
