+++ sample/template-test/stl/stub.cc
