+++ sample/template-test/stl/test.hs
